ML_Backend/info_scraper/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
firecrawl>=0.1.0
groq>=0.4.0 